Requested changes
- Prefixes to tacs_
- Cleaned up bone redefinitions
- Fixed wheel armor
- Vehicle named GT Arcadian
- Indentation fixed
- CfgVehicles moved to own file
